Add Release PORO model
This will be backed by several files rather than the database, so
doesn't extend ApplicationRecord. It represents a release of the site.
The intention is to expose this in such a way that all versions of the
frontend are available, so a specific link to the styles & javascript
bundle will be included.

Using a time-based versioning system has several benefits:

* Guaranteed uniqueness (unless releases are incredibly frequent)
* Guaranteed always-increasing version number
* Release time can be inferred from version

This separates the time parts so that the version is easier to read.