Expose Posts over GraphQL (read only)
This website will be public initially, without any interface for adding
content. I'll do all of that either later or never as I'm the only one
adding to it - I can just do it through the console. Much easier to
skip site-based permissions systems than to try and integrate it at the
moment.