Add Expense model
This models anything that cost money during the trip. Only the necessary
information is stored to recreate the CurrencyAmount object where most
of the logic belongs. The Expense describes what was bought and when -
money specifics belong in the CurrencyAmount.

Expenses can be grouped by category for a particular day as overall this
is probably more useful for comparison than each individual item.