Add data to populate the database
This is the raw data of the trip. Having it in version control means
that: a) it's easier to see what was done when, b) it's easier backup,
c) it's easier to recreate the entire database.

Running the seed task should only create the entries that do not exist,
so more data can be stored in the seeds file than is currently parsed
into database models. This can be performed later without having to
generate the data later.