Render Post content as HTML
For convenience, these are written in markdown. However, it's nicer to
show them in HTML so they can be more easily styled. It's inefficient to
do this transformation in the browser, so it's being done on the server.
The best place to do it long-term is before storing it,but this makes it
harder to edit. As I don't expect that many people to use the site,
doing the transformation at every page request isn't that bad.