Ensure releases are displayed in decscending order
For some reason this wasn't always happening, but was in the tests. This
may look like an inefficient sort, but it's actually reasonably fast to
just append .reverse:

https://stackoverflow.com/questions/2642182/sorting-an-array-in-descending-order-in-ruby

Also, this is going to be a tiny array. If it becomes slow, the
sortingcan just be done before caching. Doing it after caching means
that custom ordering could be added more easily later.