Enable Rosie to add commentary
This'll be more fun if Rosie can add her opinions & view of things
alongside mine. To do that, her contributions need to be obvious,
probably styled differently (italics for now). So this is easy to write,
I've added a custom tag into the Markdown syntax which is then replaced
with a div or span and a class. The actual tag type to replace needs
to be different depending on where text is inserted - within a
paragraph or in one by itself.

Yes, this is even more processing per page. This allows
experimetation & re-writing posts quickly, which is more valuable at
this stage. Yes, the PostType file is being poluted with renderer
modules. This will be refactored if more of them appear and / or parts
are needed elsewhere.