Add a Google Maps API key
This is hardcoded, which in general is not a good idea. However, it
can't really be avoided. It could be supplied as an environment
variable, but that would have to be compiled into the frontend bundle,
which is committed to show site progression. Therefore, it's going to
be in version control anyway. It could be supplied by an API, keeping
it out of version control, but would require an auth system and would
still be visible to the end user. The key is secured by google to only
work for specific HTTP referers, so stealing it isn't that useful.