Update site title
This is the name we've settled on. Unfortunately, in the fonts we're
using, ampersands look awful, so there needs to be a bit of a hack to
make the title one better. We could spend more time looking for a better
set of fonts, but it's not worth the time at this stage.