Make Location model more flexible
Previously, Locations were too many things:
* A point on a map
* An area on a map
* Optionally a named accommodation

These responsibilities are now more clearly split: a Location is a
named point on a map of a particular type, possibly accommodation. It is
within a named Area within a named Country. Area is more generic than
city, so can encompass towns, islands or jungle, all of which are
needed. It also makes it easier to both group and distinguish between
Locations in the same Area.