Add width, height and aspect ratio to images
Lots of frontend libraries require the dimensions and / or aspect ratio
of images. This could be calculated in javascript, but would probably be
inefficient and noticeably so. It's easier to just make this part of the
data import step, even if that does mean seeding the database requires
an internet connection.