Add a basic timeline
Initially, this is basically pagination. Eventually it will hopefully
be better, showing dates and possible more info. It uses the breakpoints
defined by bootstrap to change the size of the timeline. This allows
it to control which element is at the centre without scrolling at the
cost of not drawing all of the elements. State is deliberately stored in
the URL so pages can be exactly linked to (and state easily debugged).

Also refactors out some of the link components that are shared with the
map so that they can be reused.