Save photo sizes in YAML files
Importing photos in the seed task only worked with internet connection
as the sizes needed to be calculated. This is now moved into a new task
and the results are saved so that importing once again is independent of
internet connection (some hostels/buses/planes just don't have the
connection you'd like ...).