Start adding a colour palette
This will be useful to have for some later possibilities, eg pie charts,
but is being currently used to improve (at least slightly) the
not-so-great colours I started with for the map and the timeline and to
change up the Bootstrap navbar.

The colours are all chosen from the Material Design colour palette, so
at least sort of work together.