Use Accommodation to prevent days being duplicated
For budgeting purposes, it's easier to cinsider a day being in one
single area/country based on the location we sleep at. Without this,
using the current interface and calculations, days that involve
travelling can be double counted if areas are considered separately.
For example, on the map on the home page, a day used to be shown both
in its start area and end area - now it will only appear in the end
area.