Enable focusing on a country in the map
This is to get a rough location for a country's map, based on near where
we spent most time or between different locations. It's not perfect, but
it makes sure some areas are on screen and so is most of the country.

Also enables filtering out all of the locations not in a country to just
show what happened there.