Store a static database dump
Is this safe? Mostly. All of this information was already publicly
accessible. The site this application powers is essentially a static
site - no data changes anymore. However, it's implemented as a
Single-Page App, making use of a database and GraphQL.

Unfortunately Heroku is shutting off their free plan in a couple of
months. This app is hosted by Heroku, so I need to find another place to
host it. Storing a database dump now means that I don't _need_ to move
this off Heroku before they shut down the app - I can always spin it up
again after that point.

I may not end up needing this. The best solution here is probably to
change this to be a static site, rather than a Rails app. That way I
wouldn't have to keep anywhere near as much code up to date just to get
the app running. That may be tricky, though, and will take time. This is
a reasonable backup plan (pun semi-intended.)